+++ src/main/java/nl/tudelft/skills/controller/HomeController.java
@@ -29,6 +29,8 @@ import nl.tudelft.labracore.api.dto.*;
 import nl.tudelft.labracore.lib.security.user.AuthenticatedPerson;
 import nl.tudelft.labracore.lib.security.user.Person;
 import nl.tudelft.skills.model.SCEdition;
+import nl.tudelft.skills.model.Task;
+import nl.tudelft.skills.model.TaskCompletion;
 import nl.tudelft.skills.model.labracore.SCPerson;
 import nl.tudelft.skills.repository.EditionRepository;
 import nl.tudelft.skills.repository.ModuleRepository;
@@ -87,8 +89,9 @@ public class HomeController {
 				.filter(SCEdition::isVisible).map(SCEdition::getId).collect(Collectors.toSet());
 		Set<Long> teacherIds = person == null ? Set.of()
 				: roleApi
-						.getRolesById(editions.stream().map(EditionDetailsDTO::getId).toList(),
-								List.of(person.getId()))
+						.getRolesById(
+								editions.stream().map(EditionDetailsDTO::getId).collect(Collectors.toSet()),
+								Set.of(person.getId()))
 						.collectList().block().stream()
 						.filter(r -> r.getType() == RoleDetailsDTO.TypeEnum.TEACHER)
 						.map(r -> r.getEdition().getId()).collect(Collectors.toSet());
@@ -129,13 +132,11 @@ public class HomeController {
 			Long editionId = courseService.getLastStudentEditionForCourseOrLast(courseId);
 
 			if (editionId != null) {
-				SCEdition edition = editionService.getOrCreateSCEdition(editionId);
-				skillsDone = (int) edition.getModules().stream()
-						.flatMap(m -> m.getSubmodules().stream())
-						.flatMap(s -> s.getSkills().stream())
-						.filter(s -> s.getTasks().size() > 0
-								&& scperson.getTasksCompleted().containsAll(s.getTasks()))
-						.count();
+				List<Task> tasksDone = scperson.getTaskCompletions().stream().map(TaskCompletion::getTask)
+						.toList();
+
+				skillsDone = (int) tasksDone.stream().map(Task::getSkill).distinct()
+						.filter(s -> tasksDone.containsAll(s.getTasks())).count();
 			}
 
 			completedSkillsPerCourse.put(courseId, skillsDone);