+++ src/test/java/nl/tudelft/skills/integration/IntegrationTest.java
@@ -1,0 +1,57 @@
+/*
+ * Skill Circuits
+ * Copyright (C) 2022 - Delft University of Technology
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package nl.tudelft.skills.integration;
+
+import org.junit.jupiter.api.*;
+
+import com.microsoft.playwright.Browser;
+import com.microsoft.playwright.BrowserContext;
+import com.microsoft.playwright.Page;
+import com.microsoft.playwright.Playwright;
+
+public abstract class IntegrationTest {
+
+	private static Playwright playwright;
+	private static Browser browser;
+
+	protected BrowserContext context;
+	protected Page page;
+
+	@BeforeAll
+	static void launchBrowser() {
+		playwright = Playwright.create();
+		browser = playwright.chromium().launch();
+	}
+
+	@AfterAll
+	static void closeBrowser() {
+		playwright.close();
+	}
+
+	@BeforeEach
+	void createContextAndPage() {
+		context = browser.newContext();
+		page = context.newPage();
+	}
+
+	@AfterEach
+	void closeContext() {
+		context.close();
+	}
+
+}